Retrieve project

+++ src/Controller/IdeeController.php
@@ -19,10 +19,15 @@ class IdeeController extends Controller
      */
     public function listAction()
     {
+        $em = $this->getDoctrine()->getManager();
+
         $idees = $this->getDoctrine()
             ->getRepository(Idee::class)
             ->findBy(
-                ['user' => $this->getUser()]
+                [
+                    'user' => $this->getUser(),
+                    'user_adding' => null
+                ]
             );
 
         return $this->render('idee/list.html.twig', [
@@ -33,21 +38,31 @@ class IdeeController extends Controller
     /**
      * @Route("/idee/{id}/show", name="idee_show", requirements={"id"="\d+"})
      */
-    public function showAction($id) {
-        $comment = new Comment();
-        $formComment = $this->createForm(CommentType::class, $comment, [
-            'action' => $this->generateUrl('comment_new', ['id_idee' => $id])
-        ]);
-
+    public function showAction($id)
+    {
         $em = $this->getDoctrine()->getManager();
         $idee = $em->getRepository(Idee::class)->find($id);
 
         if (!$idee) {
             throw $this->createNotFoundException('Aucune idée trouvée pour l\'identifiant : ' . $id);
         }
+        if (!$this->checkAccessShow($idee)) {
+            return $this->redirectToRoute('idee_list');
+        }
 
+        $comment = new Comment();
+        $formComment = $this->createForm(CommentType::class, $comment, [
+            'action' => $this->generateUrl('comment_new', ['id_idee' => $id])
+        ]);
+
+        $user = $idee->getUser();
         return $this->render('idee/show.html.twig', [
             'idee' => $idee,
+            'breadcrumb' => [
+                $this->generateUrl('user_list') => "Les autres",
+                $this->generateUrl('user_show', ['id' => $user->getId()]) => $user->getFirstname(),
+                "" => $idee->getLibelle()
+            ],
             'formComment' => $formComment->createView()
         ]);
     }
@@ -61,8 +76,10 @@ class IdeeController extends Controller
 
         if (is_null($id_user)) {
             $user = $this->getUser();
+            $breadcrumb = [$this->generateUrl('idee_list') => "Mes idées", "" => "Nouvelle idée"];
         } else {
             $user = $this->getDoctrine()->getRepository(User::class)->find($id_user);
+            $breadcrumb = [$this->generateUrl('user_list') => "Les autres", "" => "Nouvelle idée pour " . $user->getFirstname()];
         }
 
         $form = $this->createForm(IdeeType::class, $idee);
@@ -81,14 +98,21 @@ class IdeeController extends Controller
             $em->persist($idee);
             $em->flush();
 
+            $this->sendEmailInsert($idee);
+
+            $this->addFlash('success', 'L\'idée a bien été ajoutée à votre liste');
+
             if (is_null($id_user)) {
+                // Si l'idée est pour le membre connecté
                 return $this->redirectToRoute('idee_list');
             } else {
+                // Si l'idée est pour un autre membre
                 return $this->redirectToRoute('user_show', ['id' => $id_user]);
             }
         }
         return $this->render('idee/new.html.twig', [
             'form' => $form->createView(),
+            'breadcrumb' => $breadcrumb,
             'user' => $user
         ]);
     }
@@ -103,18 +127,35 @@ class IdeeController extends Controller
         if (!$idee) {
             throw $this->createNotFoundException('Aucune idée trouvée pour l\'identifiant : ' . $id);
         }
+        if (!$this->checkAccessUpdate($idee)) {
+            return $this->redirectToRoute('idee_list');
+        }
 
         $form = $this->createForm(IdeeType::class, $idee);
+
+        if ($idee->getComments()) {
+            $comment = new Comment();
+            $formComment = $this->createForm(CommentType::class, $comment, [
+                'action' => $this->generateUrl('comment_new', ['id_idee' => $id])
+            ]);
+        }
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $em->flush();
+
+            // Envoi d'un mail de confirmation
+            $this->sendEmailUpdate($idee);
+
+            $this->addFlash('success', 'Les modifications ont bien été prises en compte');
             return $this->redirectToRoute('idee_list');
         }
 
         return $this->render('idee/update.html.twig', [
             'form' => $form->createView(),
+            'formComment' => $formComment->createView(),
+            'breadcrumb' => [$this->generateUrl('idee_list') => "Mes idées", "" => $idee->getLibelle()],
             'idee' => $idee
         ]);
     }
@@ -186,10 +227,69 @@ class IdeeController extends Controller
         if (!$idee) {
             throw $this->createNotFoundException('Aucun user trouvé pour l\'identifiant : ' . $id);
         }
+        if (!$this->checkAccessDelete($idee)) {
+            return $this->redirectToRoute('idee_list');
+        }
 
         $entityManager->remove($idee);
         $entityManager->flush();
 
-        return $this->redirectToRoute('idee_list');
+        $this->addFlash('success', 'L\'idée a bien été supprimée');
+
+        if (is_null($idee->getUserAdding())) {
+            // Si le membre à supprimé une de ses idées
+            return $this->redirectToRoute('idee_list');
+        } else {
+            // Si le membre a supprimé l'idée d'un autre membre
+            return $this->redirectToRoute('user_show', ['id' => $idee->getUser()->getId()]);
+        }
+    }
+
+    private function checkAccessShow(Idee $idee) {
+        return $this->getUser()->getId() != $idee->getUser()->getId();
+    }
+
+    private function checkAccessUpdate(Idee $idee) {
+        return
+            (is_null($idee->getUserAdding()) && $this->getUser()->getId() == $idee->getUser()->getId())
+            ||
+            (!is_null($idee->getUserAdding()) && $idee->getUserAdding()->getId() == $this->getUser()->getId());
+    }
+
+    private function checkAccessDelete(Idee $idee) {
+        return
+            (is_null($idee->getUserAdding()) && $this->getUser()->getId() == $idee->getUser()->getId())
+            ||
+            (!is_null($idee->getUserAdding()) && $idee->getUserAdding()->getId() == $this->getUser()->getId());
+    }
+
+    private function sendEmailInsert(Idee $idee) {
+        if ($idee->getUser()->getId() == 1) {
+            return;
+        }
+        if ($userAdding = $idee->getUserAdding()) {
+            $body = $userAdding->getFirstname() . " a ajouté l'idée " . $idee->getLibelle() . " à " . $idee->getUser()->getFirstname() . " dans votre application web !";
+        } else {
+            $body = $idee->getUser()->getFirstname() . " a ajouté l'idée " . $idee->getLibelle() . " dans votre application web !";
+        }
+        $message = (new \Swift_Message('Ajout d\'une idée cadeau'))
+            ->setFrom('nepasrepondre@loic-pascal.fr')
+            ->setTo('loic.pascal@gmail.com')
+            ->setBody(
+                $body,
+                'text/html'
+            );
+        $this->get('mailer')->send($message);
+    }
+
+    private function sendEmailUpdate(Idee $idee) {
+        $message = (new \Swift_Message('Modification d\'une idée cadeau'))
+            ->setFrom('nepasrepondre@loic-pascal.fr')
+            ->setTo('loic.pascal@gmail.com')
+            ->setBody(
+                $idee->getUser()->getFirstname() . " a mis à jour l'idée . \"" . $idee->getLibelle() . "\" dans votre application web !",
+                'text/html'
+            );
+        $this->get('mailer')->send($message);
     }
 }
